+++ src/main/java/blackjack/domain/card/Rank.java
@@ -1,33 +1,27 @@
 package blackjack.domain.card;
 
 public enum Rank {
-    ACE(1, "A"),
-    TWO(2, "2"),
-    THREE(3, "3"),
-    FOUR(4, "4"),
-    FIVE(5, "5"),
-    SIX(6, "6"),
-    SEVEN(7, "7"),
-    EIGHT(8, "8"),
-    NINE(9, "9"),
-    TEN(10, "10"),
-    KING(10, "K"),
-    QUEEN(10, "Q"),
-    JACK(10, "J");
+    ACE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    KING(10),
+    QUEEN(10),
+    JACK(10);
 
     private final int value;
-    private final String symbol;
 
-    Rank(int value, String symbol) {
+    Rank(int value) {
         this.value = value;
-        this.symbol = symbol;
     }
 
     public int getValue() {
         return value;
-    }
-
-    public String getSymbol() {
-        return symbol;
     }
 }
