+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,6 +1,6 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.card.Hand;
+import blackjack.domain.result.Hand;
 import blackjack.domain.result.Score;
 import blackjack.domain.card.Card;
 import java.util.List;
