+++ src/main/java/blackjack/domain/result/Hand.java
@@ -1,6 +1,6 @@
-package blackjack.domain.card;
+package blackjack.domain.result;
 
-import blackjack.domain.result.Score;
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.List;
