+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -3,7 +3,7 @@ package blackjack.domain.participant;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.Hand;
+import blackjack.domain.result.Hand;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.Suit;
 import blackjack.domain.result.GameResult;
