+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -3,7 +3,7 @@ package blackjack.domain.participant;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import blackjack.domain.card.Hand;
+import blackjack.domain.result.Hand;
 import org.junit.jupiter.api.Test;
 
 class PlayerTest {
